# docs/swagger.yaml
definitions:
  github_com_crazyfrankie_kube-ctl_internal_model_req.Base:
    properties:
      labels:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      name:
        type: string
      namespace:
        type: string
      restartPolicy:
        description: 'reboot strategy: Always | Never | On-Failure'
        type: string
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.ConfigMapRefVolume:
    properties:
      name:
        type: string
      optional:
        type: boolean
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.Container:
    properties:
      args:
        items:
          type: string
        type: array
      command:
        items:
          type: string
        type: array
      env:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.EnvVar'
        type: array
      envsFrom:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.EnvVarFromResource'
        type: array
      image:
        type: string
      imagePullPolicy:
        description: Always | IfNotPresent | Never
        type: string
      livenessProbe:
        allOf:
        - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ContainerProbe'
        description: Survival probes
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ContainerPort'
        type: array
      privileged:
        description: Whether to enable privileged mode (e.g. root)
        type: boolean
      readyProbe:
        allOf:
        - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ContainerProbe'
        description: Readiness Probe
      resources:
        allOf:
        - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Resource'
        description: Container application quota
      startUpProbe:
        allOf:
        - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ContainerProbe'
        description: Start the probe
      tty:
        type: boolean
      volumeMounts:
        description: Mounted volumes
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.VolumeMount'
        type: array
      workingDir:
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.ContainerPort:
    properties:
      containerPort:
        type: integer
      hostPort:
        type: integer
      name:
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.ContainerProbe:
    properties:
      command:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ProbeCommand'
      enable:
        description: Whether to turn on the probe
        type: boolean
      failureThreshold:
        description: Threshold number of probes to consider a probe failure
        type: integer
      httpGet:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ProbeHTTPGet'
      initialDelaySeconds:
        description: Initialize for a number of seconds before starting the probe
        type: integer
      periodSeconds:
        description: Probe after every few seconds
        type: integer
      successThreshold:
        description: Threshold number of probes to consider a probe successful
        type: integer
      tcpSocket:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ProbeTcpSocket'
      timeoutSeconds:
        description: Probe timeout time
        type: integer
      type:
        description: 'Probe type: tcp | http | command'
        type: string
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.DnsConfig:
    properties:
      nameservers:
        items:
          type: string
        type: array
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.DownwardAPIVolume:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.DownwardAPIVolumeItem'
        type: array
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.EnvVar:
    properties:
      name:
        type: string
      refName:
        type: string
      type:
        description: configMap | secret | default(k/v)
        type: string
      value:
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.EnvVarFromResource:
    properties:
      name:
        type: string
      prefix:
        type: string
      refType:
        description: configMap | secret
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.HostPathVolume:
    properties:
      path:
        description: host path
        type: string
      type:
        $ref: '#/definitions/v1.HostPathType'
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.Item:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.Network:
    properties:
      dnsConfig:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.DnsConfig'
      dnsPolicy:
        type: string
      hostAliases:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      hostName:
        type: string
      hostNetwork:
        type: boolean
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.NodeAffinityTermExpressions:
    properties:
      key:
        type: string
      operator:
        $ref: '#/definitions/v1.NodeSelectorOperator'
      value:
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.NodeScheduling:
    properties:
      nodeAffinity:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.NodeAffinityTermExpressions'
        type: array
      nodeName:
        type: string
      nodeSelector:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      type:
        description: nodeName | nodeSelector | nodeAffinity
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.PVCVolume:
    properties:
      claimName:
        type: string
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.Pod:
    properties:
      base:
        allOf:
        - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Base'
        description: base definition info
      containers:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Container'
        type: array
      initContainers:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Container'
        type: array
      network:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Network'
      nodeScheduling:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.NodeScheduling'
      tolerations:
        description: pod toleration params
        items:
          $ref: '#/definitions/v1.Toleration'
        type: array
      volume:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Volume'
        type: array
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.Resource:
    properties:
      CPULimit:
        type: integer
      CPUReq:
        description: cpu m
        type: integer
      enable:
        description: Whether to configure container quotas
        type: boolean
      memory:
        description: memory mi
        type: integer
      memoryLimit:
        type: integer
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.SecretRefVolume:
    properties:
      name:
        type: string
      optional:
        type: boolean
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.UpdateLabelReq:
    properties:
      labels:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      name:
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.UpdateTaintReq:
    properties:
      name:
        type: string
      taints:
        items:
          $ref: '#/definitions/v1.Taint'
        type: array
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.Volume:
    properties:
      PVCVolume:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.PVCVolume'
      configMapRefVolume:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ConfigMapRefVolume'
      downwardAPIVolume:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.DownwardAPIVolume'
      hostPathVolume:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.HostPathVolume'
      name:
        type: string
      secretRefVolume:
        $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.SecretRefVolume'
      type:
        description: emptyDir | configMap | secret | hostPath | downward | pvc
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.VolumeMount:
    properties:
      mountName:
        type: string
      mountPath:
        description: Mounted volume -> path in the corresponding container
        type: string
      readOnly:
        description: Read-only or not
        type: boolean
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_resp.Item:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_resp.Namespace:
    properties:
      createTime:
        type: integer
      name:
        type: string
      status:
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_resp.NodeDetail:
    properties:
      OSImage:
        type: string
      age:
        type: integer
      containerRuntime:
        type: string
      externalIP:
        type: string
      internalIP:
        type: string
      kernelVersion:
        type: string
      labels:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_resp.Item'
        type: array
      name:
        type: string
      status:
        type: string
      taints:
        items:
          $ref: '#/definitions/v1.Taint'
        type: array
      version:
        description: kubelet version
        type: string
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_resp.NodeListItem:
    properties:
      OSImage:
        type: string
      age:
        type: integer
      containerRuntime:
        type: string
      externalIP:
        type: string
      internalIP:
        type: string
      kernelVersion:
        type: string
      name:
        type: string
      status:
        type: string
      version:
        description: kubelet version
        type: string
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_resp.PodListItem:
    properties:
      age:
        description: Runtime
        type: integer
      ip:
        description: Pod id
        type: string
      name:
        type: string
      node:
        description: Which Node the Pod is dispatched to
        type: string
      ready:
        description: 0/1 | 1/1
        type: string
      restarts:
        description: Number of restarts
        type: integer
      status:
        description: Running | Error
        type: string
    type: object
...
  github_com_crazyfrankie_kube-ctl_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  v1.Taint:
    properties:
      effect:
        allOf:
        - $ref: '#/definitions/v1.TaintEffect'
        description: |-
          Required. The effect of the taint on pods
          that do not tolerate the taint.
          Valid effects are NoSchedule, PreferNoSchedule and NoExecute.
      key:
        description: Required. The taint key to be applied to a node.
        type: string
      timeAdded:
        description: |-
          TimeAdded represents the time at which the taint was added.
          It is only written for NoExecute taints.
          +optional
        type: string
      value:
        description: |-
          The taint value corresponding to the taint key.
          +optional
        type: string
    type: object
  v1.TaintEffect:
    enum:
    - NoSchedule
    - PreferNoSchedule
    - NoExecute
    type: string
    x-enum-varnames:
    - TaintEffectNoSchedule
    - TaintEffectPreferNoSchedule
    - TaintEffectNoExecute
  v1.Toleration:
    properties:
      effect:
        allOf:
        - $ref: '#/definitions/v1.TaintEffect'
        description: |-
          Effect indicates the taint effect to match. Empty means match all taint effects.
          When specified, allowed values are NoSchedule, PreferNoSchedule and NoExecute.
          +optional
      key:
        description: |-
          Key is the taint key that the toleration applies to. Empty means match all taint keys.
          If the key is empty, operator must be Exists; this combination means to match all values and all keys.
          +optional
        type: string
      operator:
        allOf:
        - $ref: '#/definitions/v1.TolerationOperator'
        description: |-
          Operator represents a key's relationship to the value.
          Valid operators are Exists and Equal. Defaults to Equal.
          Exists is equivalent to wildcard for value, so that a pod can
          tolerate all taints of a particular category.
          +optional
      tolerationSeconds:
        description: |-
          TolerationSeconds represents the period of time the toleration (which must be
          of effect NoExecute, otherwise this field is ignored) tolerates the taint. By default,
          it is not set, which means tolerate the taint forever (do not evict). Zero and
          negative values will be treated as 0 (evict immediately) by the system.
          +optional
        type: integer
      value:
        description: |-
          Value is the taint value the toleration matches to.
          If the operator is Exists, the value should be empty, otherwise just a regular string.
          +optional
        type: string
    type: object
  v1.TolerationOperator:
    enum:
    - Exists
    - Equal
    type: string
    x-enum-varnames:
    - TolerationOpExists
    - TolerationOpEqual
...
info:
  contact: {}
paths:
...
  /api/node:
    get:
      consumes:
      - application/json
      description: 获取集群中单个 Node 信息
      parameters:
      - description: node name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_resp.NodeDetail'
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 获取 Node 详情
      tags:
      - Node管理
  /api/node/label:
    put:
      consumes:
      - application/json
      description: 为单个 Node 添加 label
      parameters:
      - description: node name and labels
        in: body
        name: node
        required: true
        schema:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.UpdateLabelReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: Node 添加标签
      tags:
      - Node管理
  /api/node/list:
    get:
      consumes:
      - application/json
      description: 获取集群中所有 Node 信息
      parameters:
      - description: node 关键词
        in: query
        name: keyword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_resp.NodeListItem'
                  type: array
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 获取 Node 列表
      tags:
      - Node管理
  /api/node/pods:
    get:
      consumes:
      - application/json
      description: 集群中某个节点下的所有 Pod
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: Node 名称
        in: query
        name: node
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_resp.PodListItem'
                  type: array
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: Node 下的所有 Pod
      tags:
      - Node管理
  /api/node/taint:
    put:
      consumes:
      - application/json
      description: 集群中某个节点的污点更新
      parameters:
      - description: node name and taints
        in: body
        name: node
        required: true
        schema:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.UpdateTaintReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: Node 污点更新
      tags:
      - Node管理
  /api/pod:
    delete:
      consumes:
      - application/json
      description: 删除指定命名空间下的指定Pod
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: Pod名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 删除Pod
      tags:
      - Pod管理
    get:
      consumes:
      - application/json
      description: 获取指定命名空间下指定Pod的详细信息
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: Pod名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回Pod的详细信息，包含基础信息、卷配置、网络配置、初始化容器和主容器配置
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Pod'
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 获取Pod详情
      tags:
      - Pod管理
    post:
      consumes:
      - application/json
      description: 创建新的Pod或更新已存在的Pod
      parameters:
      - description: Pod配置信息
        in: body
        name: pod
        required: true
        schema:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Pod'
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功，返回成功消息
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "400":
          description: 参数错误(code=20001)或验证错误(code=20002)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 创建或更新Pod
      tags:
      - Pod管理
  /api/pod/list:
    get:
      consumes:
      - application/json
      description: 获取指定命名空间下的所有Pod列表
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回Pod列表，每个Pod包含名称、就绪状态、运行状态、重启次数、运行时长、IP和所在节点
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_resp.PodListItem'
                  type: array
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 获取Pod列表
      tags:
      - Pod管理
  /api/pod/namespace:
    get:
      consumes:
      - application/json
      description: 获取所有可用的命名空间列表
      produces:
      - application/json
      responses:
        "200":
          description: 返回命名空间列表，每个命名空间包含名称、创建时间和状态
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_resp.Namespace'
                  type: array
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 获取命名空间列表
      tags:
      - Pod管理
  /api/pod/search:
    post:
      consumes:
      - application/json
      description: 搜索指定命名空间下的指定Pod
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: Pod名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 搜索成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_resp.PodListItem'
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 搜索Pod
      tags:
      - Pod管理
...
swagger: "2.0"